+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 This projects adheres to [Semantic Versioning](https://semver.org/) and [Keep a CHANGELOG](https://keepachangelog.com/).
 
+## [13.5.0]
+
+### Fixed
+
+- Exclude provisioning fixture examples from the runtime manifest cache.
+- Rebuild warm manifest caches when the project cache version changes.
+
 ## [13.4.0]
 
 ### Added
@@ -1230,6 +1237,7 @@ Init setup
 - Gutenberg Blocks Registration.
 - Assets Manifest data.
 
+[13.4.1]: https://github.com/infinum/eightshift-libs/compare/13.4.0...13.4.1
 [13.4.0]: https://github.com/infinum/eightshift-libs/compare/13.3.0...13.4.0
 [13.3.0]: https://github.com/infinum/eightshift-libs/compare/13.2.1...13.3.0
 [13.2.1]: https://github.com/infinum/eightshift-libs/compare/13.2.0...13.2.1